Move out userController
